don't overwrite save file on failed JSON serialization;
This change creates a temp file with a datetime that is used to test whether the json serialization will succeed.I created this because sometimes the saveData had something that wasn't abled to be serialized and it would corrupt the save file.This change will let the user know if there is a problem and won't break the save file. On failure, the original save before the changes is maintained.